requirements.txt
# Google API dependencies for OAuth 2.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
requests>=2.25.0

# Image processing (usually already in ComfyUI)
Pillow>=9.0.0
numpy>=1.21.0